tests/wpt/tests/dom/events/scrolling/wheel-event-composed.html

<!DOCTYPE HTML>
<html>
<head>
<meta charset="utf-8">
<link rel="author" title="Simon Wülker" href="simon.wuelker@arcor.de">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<body>
    <script>
      promise_test(() => {
        return new Promise(resolve => {
          document.addEventListener("wheel", (event) => {
            resolve(event);
          });

          new test_driver.Actions()
            .addWheel("wheel1")
            .scroll(0, 0, 0, 10)
            .send();
        }).then((event) => {
          assert_true(event.composed);
        });
      }, `wheel event must be composed`);
    </script>
</body>

</html>
